requirements.txt
absl-py==0.9.0
albumentations==0.4.5
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
efficientnet==1.0.0
efficientnet-pytorch==0.6.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
grpcio==1.30.0
h5py==2.10.0
idna==2.9
image-classifiers==1.0.0
imageio==2.8.0
imgaug==0.2.6
importlib-metadata==1.6.1
Keras-Applications==1.0.8
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.2
munch==2.5.0
networkx==2.4
numpy==1.19.0
oauthlib==3.1.0
opencv-python-headless==4.2.0.34
pandas==1.0.5
Pillow==7.1.2
pretrainedmodels==0.7.4
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.0
segmentation-models-pytorch==0.1.0
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tifffile==2020.6.3
torch==1.3.1
torchvision==0.4.2
tqdm==4.46.1
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
